pyproject.toml
[project]
name = "quant-factor-portfolio-pipeline"
version = "0.1.0"
description = "End-to-end quantitative factor investment pipeline: from alpha discovery to portfolio deployment"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26,<2.1",
    "pandas>=2.2,<2.3",
    "scipy>=1.13,<1.15",
    "statsmodels>=0.14",
    "scikit-learn>=1.3",
    "cvxpy>=1.4",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pyyaml>=6.0",
    "openpyxl>=3.1",
    "pyarrow>=12.0",
    "tqdm>=4.65",
    "plotly",
    "kaleido",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov", "ruff", "mypy"]
notebook = ["jupyterlab>=4.0", "ipywidgets"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
